nouveau système de MapXML fini et complètement fonctionnel
